@@ -15,7 +15,7 @@
     "clean-webpack-plugin": "3.0.0",
     "metadata-filter": "0.4.1",
     "simple-icons": "2.18.0",
-    "vue": "2.6.11",
+    "vue": "3.0.0-rc.7",
     "webextension-polyfill-ts": "0.19.0"
   },
   "devDependencies": {
@@ -24,6 +24,7 @@
     "@types/chai": "4.2.12",
     "@types/chai-as-promised": "7.1.3",
     "@types/mocha": "8.0.0",
+    "@vue/compiler-sfc": "3.0.0-rc.7",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
     "chai-spies": "1.0.0",
@@ -71,9 +72,8 @@
     "ts-node": "8.10.2",
     "tsconfig-paths-webpack-plugin": "3.2.0",
     "typescript": "4.0.2",
-    "vue-loader": "15.9.2",
+    "vue-loader": "16.0.0-beta.5",
     "vue-style-loader": "4.1.2",
-    "vue-template-compiler": "2.6.11",
     "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.11.0"
