@@ -1,6 +1,8 @@
-import Vue, { VueConstructor } from 'vue';
+import { createApp, Component } from 'vue';
 
 import { L } from '@/common/i18n';
+
+const rootElementSelector = '#content';
 
 const i18nMixin = {
 	methods: { L },
@@ -14,16 +16,14 @@ const removeSvgTitleMixin = {
 	},
 };
 
-Vue.mixin(i18nMixin);
-Vue.mixin(removeSvgTitleMixin);
-
 /**
  * Create a new instance of a Vue model.
  *
  * @param component Vue component instance
  */
-export function createApp(component: VueConstructor<Vue>): void {
-	new Vue({
-		render: (h) => h(component),
-	}).$mount('#content');
+export function createVueApp(component: Component): void {
+	createApp(component)
+		.mixin(i18nMixin)
+		.mixin(removeSvgTitleMixin)
+		.mount(rootElementSelector);
 }
