@@ -3,6 +3,6 @@ import 'bootstrap/dist/css/bootstrap.css';
 import '@/ui/base.css';
 
 import root from '@/ui/options/root.vue';
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
-createApp(root);
+createVueApp(root);
