@@ -1,8 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import '@/ui/base.css';
 
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
 import Disabled from '@/ui/popups/disabled.vue';
 
-createApp(Disabled);
+createVueApp(Disabled);
