@@ -1,8 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import '@/ui/base.css';
 
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
 import _Error from '@/ui/popups/error.vue';
 
-createApp(_Error);
+createVueApp(_Error);
