@@ -1,8 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import '@/ui/base.css';
 
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
 import GoPlayMusic from '@/ui/popups/go-play-music.vue';
 
-createApp(GoPlayMusic);
+createVueApp(GoPlayMusic);
