@@ -1,5 +1,5 @@
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
 import Info from '@/ui/popups/info.vue';
 
-createApp(Info);
+createVueApp(Info);
