@@ -1,8 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import '@/ui/base.css';
 
-import { createApp } from '@/ui/vue-helper';
+import { createVueApp } from '@/ui/vue-helper';
 
 import Unsupported from '@/ui/popups/unsupported.vue';
 
-createApp(Unsupported);
+createVueApp(Unsupported);
